# [Storage] Optimize speed & memory of indexing collection (BadgerDB, Pebble)