--- a/storage/store/collection_bench_test.go
+++ b/storage/store/collection_bench_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/dgraph-io/badger/v2"
+
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/module/metrics"
 	"github.com/onflow/flow-go/storage/operation/badgerimpl"