Currently, during collection indexing, transaction fingerprint is computed twice for every transaction in the collection:
- creating light collection
- indexing transaction

This PR optimizes collection indexing by reusing the transaction ID:
- add `Transaction.StoreByID()` to index transaction by given fingerprint
- update `Collections.Store()` to reuse transaction fingerprint
- update `IndexerCore.IndexBlockData()` to reuse transaction fingerprint
- etc.

I found this while getting familiar with storage code related to BadgerDB & Pebble (not looking for optimizations but it saves time to wrap up while details are in mind).

### Benchmarks using BadgerDB

Benchmarks for indexing collection containing 5 transactions:

```
                             │ original.log │            optimized.log             │
                             │    sec/op    │    sec/op     vs base                │
IndexCollectionWithBadger-12   69.20µ ± 15%   60.81µ ± 15%  -12.12% (p=0.010 n=20)

                             │ original.log │            optimized.log             │
                             │     B/op     │     B/op      vs base                │
IndexCollectionWithBadger-12   43.28Ki ± 4%   37.32Ki ± 4%  -13.76% (p=0.000 n=20)

                             │ original.log │           optimized.log            │
                             │  allocs/op   │ allocs/op   vs base                │
IndexCollectionWithBadger-12     368.0 ± 0%   303.0 ± 0%  -17.66% (p=0.000 n=20)
```