@fxamacker 2025-04-15
> IMO, we could just leave this to the Malleability task force. Once https://github.com/onflow/flow-go/issues/7303, Store(tx) should be able to safely reuse the cached ID.

Thanks for the context!

Closing this PR.